Make HostedBranch a class
