+++ test/src/main/java/org/openjdk/skara/test/TestHostedRepository.java
@@ -190,7 +190,7 @@ public class TestHostedRepository extends TestIssueProject implements HostedRepo
         try {
             var result = new ArrayList<HostedBranch>();
             for (var b : localRepository.branches()) {
-                result.add(new TestHostedBranch(b.name(), localRepository.resolve(b).orElseThrow()));
+                result.add(new HostedBranch(b.name(), localRepository.resolve(b).orElseThrow()));
             }
             return result;
         } catch (IOException e) {
